Improve implementation of `git changelog-from-parent`
Now it optionally accepts a parameter

Prints which diff the command is trying to calculate

Only tries to deduce the parent branch if the parameter was not
specified.

The main issue is that `deduce_parent_branch` is not quite stable just
yet, i need to find a better way to implement this in a reliable way,
but for now i will make it easier to work with this command.